refactor: move RegisterHelper to utils
